RAG/requirements.txt
streamlit==1.18.0
python-dotenv==1.0.0
langchain==0.0.16
langchain-community==0.0.4
langchain-text-splitters==0.0.1
langchain-groq==0.0.1
faiss-cpu==1.7.3
huggingface-hub==0.12.1
